EMBER3D version that can be run on CPU
The old version requires GPU for running. This runs on CPU. However, it is still too costly to run on surface locally. To run this, you need to use a Conda environment on python 3.9. Also, the DGL library needs to be installed separately because the requirements.txt file is out of date.